requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
Django==3.2.7
django-cors-headers==3.8.0
django-dynamic-raw-id==2.8
django-filter==2.4.0
django-grappelli==2.15.1
django-modeladmin-reorder==0.3.1
djangorestframework==3.12.4
geoip2==4.3.0
idna==3.2
Markdown==3.3.4
maxminddb==2.1.0
multidict==5.1.0
numpy==1.21.2
packaging==21.0
pandas==1.3.3
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.6
xlrd==2.0.1
yarl==1.6.3
